Limit purpose of `PathModification` and rename it
It is now renamed to `PathBase` and only controls base paths of printed
paths.

This is because the meaning of "path modification" is too wide.
`--workdir` option (whether to show working directory or repository
itself) can be thought as "path modification", but I didn't intend to
represent such options by `PathModification`.