Skip traversal only when the directory is actually workdir
Previous implementation always skips rest descendants of `foo/` when
`foo/.git` is a git repository.
However, it is not sure whether the parent of `.git` is really workdir,
because the working directory of a git repository can be changed by
repository-local config.

This commit makes sure that the directory containing `.git` is skipped
only when it is really workdir.